requirements.txt
altgraph==0.17.2
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
django-filter==23.3
django-mathfilters==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
future==0.18.2
gunicorn==20.1.0
idna==3.4
install==1.3.5
macholib==1.15.2
oauthlib==3.2.2
pip==21.2.4
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==58.0.4
six==1.15.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==1.26.16
wheel==0.37.0
whitenoise==6.5.0
